Restructure app in preperation for docker
