+++ app/forecasts.txt
@@ -1,0 +1,1 @@
+20180408-130000
