--- forecasts.txt
@@ -1,1 +1,0 @@
-20180408-130000
