+++ front-end/webpack.dev.js
@@ -7,7 +7,7 @@ module.exports = merge(common, {
     contentBase: "./dist",
     proxy: {
       "/api/*": {
-        target: "http://localhost:8000",
+        target: "http://localhost:5000",
         secure: false
       }
     },
